+++ packages/react/src/components/Field.tsx
@@ -140,6 +140,7 @@ export function Field<
                 // state also resets to "false" when user removes input
                 effect(() => {
                   if (
+                    element.type !== 'radio' &&
                     field.startValue.value === undefined &&
                     field.value.peek() === undefined
                   ) {