+++ packages/preact/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Refactor `getFilteredNames` util
+
 ## v0.5.2 (June 17, 2023)
 
 - Allow `noValidate` prop of `Form` to be overwritten (issue #92)