+++ packages/preact/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## v0.5.4 (June 25, 2023)
+## vX.X.X (Month DD, YYYY)
 
 - Fix date support for `insert` and `replace` method