+++ packages/preact/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v0.6.1 (July 31, 2023)
 
 - Fix `valiField` and `valiForm` adapter for Valibot v0.7.0